Split llm tools into focused modules
